WoT restructure: rename utils/hitTest.ts -> utils/pointerToCoords.ts (module now maps pointer position -> grid coords)
