+++ handsontable/src/3rdparty/walkontable/src/utils/pointerToCoords.ts
